refactor: move surviving compiler primitives
