--- src/functions.ts
@@ -1,7 +1,0 @@
-export type {
-  FunctionName,
-  FunctionOutputName,
-  FunctionReturnType,
-  FunctionReturnTypes,
-  IsFunctionCall,
-} from './compiler/semantics/functions.js';
