+++ src/public/schema.ts
@@ -5,7 +5,7 @@ export type {
   SchemaLike,
 } from '../compiler/schema/model.js';
 
-export type { FunctionReturnTypes } from '../compiler/legacy.js';
+export type { FunctionReturnTypes } from '../compiler/semantics/functions.js';
 
 export type { QueryTypeError } from '../compiler/contracts/public-error.js';
